+++ Server/src/GameWorld/Player/Player.h
@@ -33,7 +33,7 @@ class Player {
         void setShootingState();
         void stopShootingState();
         void applyStep(std::map<int, std::shared_ptr<Collidable>>& collidables,
-                       std::map<uint8_t, std::shared_ptr<Infected>>& infecteds);
+                       std::map<int, std::shared_ptr<Infected>>& infecteds);
         std::pair<int16_t, int16_t>& getPosition();
         TypeOperator& getTypeOperator();
         State& getState();
@@ -41,7 +41,7 @@ class Player {
         virtual ~Player() = default;
     private:
         void move(std::map<int, std::shared_ptr<Collidable>>& collidables);
-        void shoot(std::map<uint8_t, std::shared_ptr<Infected>>& infecteds);
+        void shoot(std::map<int, std::shared_ptr<Infected>>& infecteds);
         virtual void specialAtack(Event event) = 0;
 };
